fix bug that broke ticket validation
This is an API change, so it really should be a minor version update,
not a patch, but I just published the package like an hour ago and I'm
confident no one's using it, so I'm violating semver just this once.

This is what happens when you don't have automated testing. I should
(and do) feel bad.

Anyway, what happened was that serviceUrl was supposed to be included in
the validation request as well as the login request, and I didn't
realize that. So I wrote this big bad bug that broke ticket validation
for versions 1.0, 2.0, and 3.0 -- ironically, everything but SAML.